// src/handlers/sns/fxtweet.d.ts
export interface TweetAPIResponse {
  code: number;
  message: string;
  tweet?: APITwitterStatus;
}

interface StatusAPIResponse {
  code: number;
  message: string;
  status?: APITwitterStatus;
}

interface UserAPIResponse {
  code: number;
  message: string;
  user?: APIUser;
}

interface APITranslate {
  text: string;
  source_lang: string;
  source_lang_en: string;
  target_lang: string;
}

interface APIExternalMedia {
  type: "video";
  url: string;
  thumbnail_url?: string;
  height?: number;
  width?: number;
}

interface APIPollChoice {
  label: string;
  count: number;
  percentage: number;
}

interface APIPoll {
  choices: APIPollChoice[];
  total_votes: number;
  ends_at: string;
  time_left_en: string;
}

interface APIMedia {
  type: string;
  url: string;
  width: number;
  height: number;
}

interface APIPhoto extends APIMedia {
  type: "photo";
  altText: string;
}

interface APIVideo extends APIMedia {
  type: "video" | "gif";
  thumbnail_url: string;
  format: string;
  duration: number;
  variants: TweetMediaFormat[];
}

interface APIMosaicPhoto extends APIMedia {
  type: "mosaic_photo";
  formats: {
    webp: string;
    jpeg: string;
  };
}

interface APIStatus {
  id: string;
  url: string;
  text: string;
  created_at: string;
  created_timestamp: number;

  likes: number;
  reposts: number;
  replies: number;

  quote?: APIStatus;
  poll?: APIPoll;
  author: APIUser;

  media: {
    external?: APIExternalMedia;
    photos?: APIPhoto[];
    videos?: APIVideo[];
    all?: APIMedia[];
    mosaic?: APIMosaicPhoto;
  };

  lang: string | null;
  possibly_sensitive: boolean;

  replying_to: {
    screen_name: string;
    post: string;
  } | null;

  source: string | null;

  embed_card: "tweet" | "summary" | "summary_large_image" | "player";
  provider: DataProvider;
}

interface APITwitterCommunityNote {
  text: string;
  entities: BirdwatchEntity[];
}

interface APITwitterStatus extends APIStatus {
  views?: number | null;
  translation?: APITranslate;

  is_note_tweet: boolean;
  community_note: APITwitterCommunityNote | null;
  provider: DataProvider.Twitter;
}